requirements.txt
colorama==0.4.6
iniconfig==2.0.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
